Replaced all placeholder calls to beans
Web resources now parse JSON messages to use beans dynamically

Used jackson-databind api to convert incoming JSON messages into a tree.
Parsing consists of tree-search for a specific key.  Nodes with the
specified keys are parsed for their value.

Presently, the web resources are expecting very specific keys to be
passed in the messages, with no error handling in the case of invalid
keys.

Added an entry into the Maven dependencies for Jackson libraries

Updated methods in school bean to expose more robust invocation
parameters.  However, it may be desirable to aggregate the parameters
into a structure.