Renamed packages, new web-oriented package for client, new project for a shared library between client and server
Package names all follow the pattern shield.<project>.<package>

New project to group code to be shared between both client and server,
such as the intermediate representation of objects transmitted between
client and server.

New package in client to consolidate code for communicating with server.
Presently only groups URLS, in the future the controller classes will
use it to do all the communication.