Added function to add schools to database via client-side GUI
GUI sends a POST message to appropriate web resource with a JSON payload
that is translated into a simple Java POJO for a school.  The properties
of the POJO are remapped into a School entity and passed to the EJB.
The EJB persists the School to the database.

Commented out some explicit db transaction creation code.

Renamed a field in the client's fxml files to correspond with the the
new package path name.