Friendship entity connects two Students.  Merge functionality of FriendRequest* and Friend* classes.
Friendship is a managed entity associating two students.  More efficient
representation in database than lists of friends for each student.
Each Friendship object is unique up to the combination of its sender and
recipient.

An approved flag indicates whether the Friendship is just a friend
request or if has been accepted.  Accepting a friend request approves
the friendship object and automatically creates and approves a
friendship in the opposite direction (sender and recipient inverted).

This ensures the symmetry of the friendship graph, and prevents Students
from creating a new friendship with any other Student they are friends
with, and not just with Students they have already sent a friend request
to.

Since the Friendship entity unites the notions of friends and friend
requests, the corresponding beans and web resources have also been
united.

Still @TODO, deleting friends.