Converted static ServerResource strings to static ServerResource objects.  ServerAccessPoints now only accept ServerResource objects.
Should not affect any pre-existing functionality.  The only intended
effect of this change is you can't create a ServerAccessPoint for a
client screen using a string or url that isn't defined in the
ServerResource class.  The ServerResource class has a private
constructor.  To create a new ServerAccessPoint you must use one of the
ServerResource objects that are defined in the ServerResource class.