Set up design of schedule block class, added some constraints to properties of Schools
Schedule blocks are unique up to the combination of their School,
period, and the set of days they use.

Since each day in the set is unique, the set of days is internally
encoded as a String.  E.g. a schedule block using days 1,2, and 4 will
represent its days as "124."  This allows us to store the set of days as
a basic value in a single column to maintain the uniqueness constraint
detailed above.

Added minimum and maximum value annotations to the properties of School
entities.