Cleaned up Course design, adding courses is done through a School object
Courses have a package private constructor.  The effect of this and the
previous commit is that we can't persist an orphaned Course or an
orphaned Section.  Courses only get persisted by being added to the
School's list, and Sections only by being added to the Course's list.

Fixed uniqueness constraint for Courses.  Courses are now unique up to
the combination of their SCHOOL and their Identifier.  This way two
schools can have "PHY101" for example.

Fixed a bug in the delcaration of the unique constraints for Friendship,
and ScheduleBlock.