Implemented Generate Desired Schedule algorithm; Creating a new School creates a Lunch class for each day with Schedule Blocks and Sections
Generate Schedule recursive backtracking algorithm.
Tries to add a not-excluded section for each course desired.
Returns early if two or more courses couldn't fit into the schedule.
Returns a list of perfectly fit schedules or a list of one-off
schedules, both sorted by overlap with friends' schedules.

Lunches are treated internally as a unique course for each day in the
school week, with a section for every period they are offered.  Adding a
school generates these lunches (in the AdminSchoolBean), creates
Schedule Blocks for each section, and adds the sections.  The
ScheduleBlocks are flagged as special lunch schedule blocks to
distinguish them from normal ScheduleBlocks.

The gross effect of this is to simplify the design of the backtracking
algorithm.  Lunches are added or omitted just like every other course
(preventing collisions).

@TODO add special omission counter just for lunches