Generate Schedule algorithm treats criteria omissions per requirements
Specifically, the omission of a course is equivalent to omitting lunch
on every day of the schedule-week.

The algorithm tracks an omission score during the search through
schedule space.  Omitting a lunch on one day is worth 1 point.  Omitting
a course is worth d points, where d is the number of days in the
schedule-week.  If the search reaches an omission score of more than d
points then it short-circuits that branch of the search path.