Resolved bug in Course Collection objects for School entity.  getCourses() returns all non-lunch courses.  getLunches() returns all lunches.
Since both Collections objects were OneToMany collections of Courses,
and since the Course entity must be the owner of the bidirectional
relationship with School entities (to enforce unique constraint on
Course table), the collections were being reconstructed from the
database using the same query.  Consequently, both collections were the
same list of courses, i.e. all of them.

Solution for now is to sift through the courses and remove the type of
courses we don't need for that operation.  E.g. for getLunches() we
filter out courses that aren't lunches.  Penalty is O(n) on the size of
all courses.

Alternative solution is to have a separate entity for Lunches, sharing a
type with Course (probably a sub-class).